working code, but can't linked

+++ mm/set_use_memory.c
@@ -1,8 +1,9 @@
-#include "memory_struct.h"
+#include <memory_struct.h>
 #include <kstdint.h>
 
-uint8_t set_use_memory(uint64_t size)
+uint64_t set_use_memory(uint64_t size)
 {
+    
     
     return 0;
 }
